test(server): move live reload tests to new file
