+++ test/ContentBase.test.js
@@ -18,109 +18,6 @@ const contentBaseOther = path.join(
 describe('ContentBase', () => {
   let server;
   let req;
-  describe('Test disabling live reloading', () => {
-    const nestedFile = path.join(contentBasePublic, 'assets/example.txt');
-
-    jest.setTimeout(30000);
-
-    beforeAll((done) => {
-      server = helper.start(
-        config,
-        {
-          contentBase: contentBasePublic,
-          watchContentBase: true,
-          liveReload: false,
-        },
-        done
-      );
-      req = request(server.app);
-    });
-
-    afterAll((done) => {
-      helper.close(() => {
-        done();
-      });
-      fs.truncateSync(nestedFile);
-    });
-
-    it('Should not reload on changing files', (done) => {
-      let reloaded = false;
-
-      server.contentBaseWatchers[0].on('change', () => {
-        // it means that file has changed
-
-        // simulating server behaviour
-        if (server.options.liveReload !== false) {
-          Object.defineProperty(window.location, 'reload', {
-            configurable: true,
-          });
-          window.location.reload = jest.fn();
-          window.location.reload();
-          reloaded = true;
-        }
-        expect(reloaded).toBe(false);
-
-        done();
-      });
-
-      // change file content
-      setTimeout(() => {
-        fs.writeFileSync(nestedFile, 'Heyo', 'utf8');
-      }, 1000);
-    });
-  });
-
-  describe('Testing live reloading', () => {
-    const nestedFile = path.join(contentBasePublic, 'assets/example.txt');
-
-    jest.setTimeout(30000);
-
-    beforeAll((done) => {
-      server = helper.start(
-        config,
-        {
-          contentBase: contentBasePublic,
-          watchContentBase: true,
-          liveReload: true,
-        },
-        done
-      );
-      req = request(server.app);
-    });
-
-    afterAll((done) => {
-      helper.close(() => {
-        done();
-      });
-      fs.truncateSync(nestedFile);
-    });
-
-    it('Should reload on changing files', (done) => {
-      let reloaded = false;
-
-      server.contentBaseWatchers[0].on('change', () => {
-        // it means that files has changed
-
-        // simulating server behaviour
-        if (server.options.liveReload !== false) {
-          Object.defineProperty(window.location, 'reload', {
-            configurable: true,
-          });
-          window.location.reload = jest.fn();
-          window.location.reload();
-          reloaded = true;
-        }
-        expect(reloaded).toBe(true);
-
-        done();
-      });
-
-      // change file content
-      setTimeout(() => {
-        fs.writeFileSync(nestedFile, 'Heyo', 'utf8');
-      }, 1000);
-    });
-  });
 
   describe('to directory', () => {
     const nestedFile = path.join(contentBasePublic, 'assets/example.txt');
